Add initial code
Stable releases list looks good. Next step is the prereleases
list.

Then it should just be creating the feed and setting up the
Action?

Probably should set it to run daily, but only commit to the
repo if the feed changed, or once every 2wks or so.

But, commits are cheap; might not be worth the trouble.